Moved components to own folders
