+++ client/src/components/Modal/Modal.tsx
@@ -1,6 +1,6 @@
 import { MouseEvent, useState } from 'react';
-import LoginForm from './LoginForm';
-import RegisterForm from './RegisterForm';
+import LoginForm from '../LoginForm/LoginForm';
+import RegisterForm from '../RegisterForm/RegisterForm';
 
 type Props = {
   toggleModal: () => void;
