+++ client/src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App.tsx';
-import './index.css';
+//import './index.css';
 import CustomerContextProvider from './context/CustomerContext.tsx';
 import CartContextProvider from './context/CartContext.tsx';
