+++ client/src/pages/StartPage/StartPage.tsx
@@ -1,4 +1,4 @@
-import ProductList from '../components/ProductList';
+import ProductList from '../../components/ProductList/ProductList';
 
 type Props = {};
